collect_my_payout_docs/requirements.txt
python-dotenv==1.1.0
pdfplumber==0.11.6
openai==1.76.0
PyYAML==6.0.2
tqdm==4.66.4
stripe==7.11.0
pypdf==4.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
